services/notification/requirements.txt
fastapi
uvicorn
python-dotenv
pytest
pydantic[email]